Add files via upload
Implemented a dark/light mode. Always redid how the videos were shown after clicking a character. Instead of it being a dropdown it now plays as a popup in the middle of the screen. I also made the videos loop so users wont have to replay it manually. CSS is updated when necessary